tech(api): remove unused code (old chat and message models + repo method)
