+++ api/tests/devcomp/acceptance/application/passages/passage-controller_test.js
@@ -383,7 +383,6 @@ describe('Acceptance | Controller | passage-controller', function () {
                 context: 'some context',
               },
             }),
-            hasAttachmentContextBeenAdded: false,
             messages: [],
           });
           const chatDTO = chat.toDTO();
