+++ api/tests/evaluation/acceptance/application/assessments/assessment-controller_test.js
@@ -869,7 +869,6 @@ describe('Acceptance | Controller | assessment-controller', function () {
                 context: 'some context',
               },
             }),
-            hasAttachmentContextBeenAdded: false,
             messages: [],
           });
           const chatDTO = chat.toDTO();
