+++ api/tests/llm/integration/domain/usecases/start-chat_test.js
@@ -1,4 +1,4 @@
-import { ChatV2 } from '../../../../../src/llm/domain/models/ChatV2.js';
+import { Chat } from '../../../../../src/llm/domain/models/Chat.js';
 import { Configuration } from '../../../../../src/llm/domain/models/Configuration.js';
 import { startChat } from '../../../../../src/llm/domain/usecases/start-chat.js';
 import { chatRepository, configurationRepository } from '../../../../../src/llm/infrastructure/repositories/index.js';
@@ -34,7 +34,7 @@ describe('LLM | Integration | Domain | UseCases | start-chat', function () {
 
         // then
         expect(chat).to.deepEqualInstance(
-          new ChatV2({
+          new Chat({
             id: '123e4567-e89b-12d3-a456-426614174000',
             configuration: new Configuration({}), // Configuration’s properties are not enumerable
             totalInputTokens: 0,
@@ -83,7 +83,7 @@ describe('LLM | Integration | Domain | UseCases | start-chat', function () {
 
         // then
         expect(chat).to.deepEqualInstance(
-          new ChatV2({
+          new Chat({
             id: '123e4567-e89b-12d3-a456-426614174000',
             userId: 123456,
             assessmentId: 11,
