+++ api/tests/llm/unit/domain/models/Chat_test.js
@@ -1,37 +1,66 @@
+import {
+  IncorrectMessagesOrderingError,
+  MaxPromptsReachedError,
+  NoAttachmentNeededError,
+  NoAttachmentNorMessageProvidedError,
+  TooLargeMessageInputError,
+} from '../../../../../src/llm/domain/errors.js';
 import { Chat, Message } from '../../../../../src/llm/domain/models/Chat.js';
 import { Configuration } from '../../../../../src/llm/domain/models/Configuration.js';
-import { expect } from '../../../../test-helper.js';
+import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('LLM | Unit | Domain | Models | Chat', function () {
-  describe('#get isPreview', function () {
-    context('when chat is not linked to a user', function () {
-      it('should return true', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [],
-        });
+  describe('#get hasAttachmentContextBeenAdded', function () {
+    context('when there are no user messages with the expected attachment', function () {
+      it('returns false', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Peux tu m aider ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'oui bien sur' }),
+          domainBuilder.llm.buildUserMessage({
+            index: 2,
+            content: 'que contient cette piece jointe ?',
+            attachmentName: 'wrong_file.txt',
+          }),
+          domainBuilder.llm.buildAssistantMessage({ index: 3, content: 'je sais pas ce que c' }),
+        ];
+        const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'right_file.txt' });
+        const chat = domainBuilder.llm.buildChat({ messages, configuration });
 
-        // then
-        expect(chat).to.have.property('isPreview', true);
+        expect(chat.hasAttachmentContextBeenAdded).to.be.false;
       });
     });
 
-    context('when chat is linked to a user', function () {
-      it('should return false', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          userId: 123,
-          configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [],
-        });
+    context('when there is a user message with the expected attachment', function () {
+      it('returns true', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Peux tu m aider ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'oui bien sur' }),
+          domainBuilder.llm.buildUserMessage({
+            index: 2,
+            content: 'que contient cette piece jointe ?',
+            attachmentName: 'right_file.txt',
+          }),
+          domainBuilder.llm.buildAssistantMessage({ index: 3, content: 'je sais pas ce que c' }),
+        ];
+        const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'right_file.txt' });
+        const chat = domainBuilder.llm.buildChat({ messages, configuration });
 
-        // then
-        expect(chat).to.have.property('isPreview', false);
+        expect(chat.hasAttachmentContextBeenAdded).to.be.true;
+      });
+    });
+
+    context('when configuration does not expect an attachment to be transmitted', function () {
+      it('throws', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Peux tu m aider ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'oui bien sur' }),
+        ];
+        const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: null });
+        const chat = domainBuilder.llm.buildChat({ messages, configuration });
+
+        expect(() => chat.hasAttachmentContextBeenAdded).to.throw(
+          'Configuration must have an attachment config setup to call this getter',
+        );
       });
     });
   });
@@ -44,7 +73,6 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
           configuration: new Configuration({
             challenge: { victoryConditions: { expectations: ['super_victory_condition'] } },
           }),
-          hasAttachmentContextBeenAdded: false,
           messages: [],
         });
 
@@ -58,7 +86,6 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
         const chat = new Chat({
           id: 'some-chat-id',
           configuration: new Configuration({}),
-          hasAttachmentContextBeenAdded: false,
           messages: [],
         });
 
@@ -68,1079 +95,956 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
     });
   });
 
-  describe('#addUserMessage', function () {
-    it('should append a message as user message when message has content', function () {
-      // given
-      const message = 'un message pas vide';
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [new Message({ index: 0, content: 'some message', isFromUser: false })],
-      });
+  describe('#get isPreview', function () {
+    context('when userId is defined', function () {
+      it('returns false', function () {
+        const chat = domainBuilder.llm.buildChat({ userId: 789 });
 
-      // when
-      chat.addUserMessage(message, false, false, false, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some message',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: false,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-        {
-          index: 1,
-          content: 'un message pas vide',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: true,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: false,
-          wasModerated: false,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should not append anything when message has no content', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [new Message({ index: 0, content: 'some message', isFromUser: false })],
-      });
-
-      // when
-      chat.addUserMessage('', false, false, false, false);
-      chat.addUserMessage(null, false, false, false, false);
-      chat.addUserMessage(undefined, false, false, false, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some message',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: false,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set shouldBeCountedAsAPrompt at true if given true', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addUserMessage('some content', true, false, false, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some content',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: true,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: true,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: false,
-          wasModerated: false,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set shouldBeForwardedToLLM at true if given true', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addUserMessage('some content', false, true, false, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some content',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: true,
-          shouldBeForwardedToLLM: true,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: false,
-          wasModerated: false,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set haveVictoryConditionsBeenFulfilled at true if given as true', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addUserMessage('some content', false, false, true, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some content',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: true,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: true,
-          wasModerated: false,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set wasModerated at true if given as true', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addUserMessage('some content', false, false, false, true);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some content',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: true,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: false,
-          wasModerated: true,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-  });
-
-  describe('#addLLMMessage', function () {
-    it('should append a message as llm message when message has content', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [new Message({ index: 0, content: 'some message', isFromUser: false })],
-      });
-
-      // when
-      chat.addLLMMessage('un message pas vide', true);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some message',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: false,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-        {
-          index: 1,
-          content: 'un message pas vide',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: true,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should not append anything when message has no content', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [new Message({ index: 0, content: 'some message', isFromUser: false })],
-      });
-
-      // when
-      chat.addLLMMessage('', true);
-      chat.addLLMMessage(null, true);
-      chat.addLLMMessage(undefined, true);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'some message',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: false,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set "shouldBeForwardedToLLM" to the value passed in parameter', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addLLMMessage('message1', true);
-      chat.addLLMMessage('message2', false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'message1',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: true,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-        {
-          index: 1,
-          content: 'message2',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: false,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: undefined,
-        },
-      ]);
-    });
-
-    it('should set "hasErrorOccurred" to the value passed in parameter', function () {
-      // given
-      const chat = new Chat({
-        id: 'some-chat-id',
-        configuration: new Configuration({ id: 'some-config-id' }),
-        hasAttachmentContextBeenAdded: false,
-        messages: [],
-      });
-
-      // when
-      chat.addLLMMessage('message1', true, true);
-      chat.addLLMMessage('message2', true, false);
-
-      // then
-      expect(chat.toDTO()).to.have.deep.property('messages', [
-        {
-          index: 0,
-          content: 'message1',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: true,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: true,
-        },
-        {
-          index: 1,
-          content: 'message2',
-          attachmentName: undefined,
-          attachmentContext: undefined,
-          isFromUser: false,
-          shouldBeForwardedToLLM: true,
-          shouldBeRenderedInPreview: true,
-          shouldBeCountedAsAPrompt: false,
-          hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-          haveVictoryConditionsBeenFulfilled: undefined,
-          wasModerated: undefined,
-          hasErrorOccurred: false,
-        },
-      ]);
-    });
-  });
-
-  describe('#addAttachmentContextMessages', function () {
-    context('when attachment context has not been added yet', function () {
-      context('when attachment is the right one', function () {
-        context('when attachment has been submitted along with a message', function () {
-          it('should add two fictional messages, one from the user and the other one from the LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: false,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('winter_lyrics.txt', 'some user message along with the attachment');
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: true,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: true,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 3,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                isFromUser: false,
-                shouldBeForwardedToLLM: true,
-                shouldBeRenderedInPreview: false,
-                shouldBeCountedAsAPrompt: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-        context('when attachment has been submitted alone', function () {
-          it('should add two fictional messages, one from the user and the other one from the LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: false,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('winter_lyrics.txt', null);
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: true,
-                shouldBeForwardedToLLM: true,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 3,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                isFromUser: false,
-                shouldBeForwardedToLLM: true,
-                shouldBeRenderedInPreview: false,
-                shouldBeCountedAsAPrompt: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-      });
-
-      context('when attachment is the wrong one', function () {
-        context('when attachment has been submitted along with a message', function () {
-          it('should add a user fictional message and not add any assistant context message, and should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: false,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('summer_lyrics.txt', 'some user message along with the attachment');
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', false);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'summer_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: true,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-        context('when attachment has been submitted alone', function () {
-          it('should add a user fictional message and not add any assistant context message, and should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: false,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('summer_lyrics.txt', null);
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', false);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'summer_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: true,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
+        expect(chat.isPreview).to.be.false;
       });
     });
 
-    context('when attachment context has already been added', function () {
-      context('when attachment is the right one', function () {
-        context('when attachment has been submitted along with a message', function () {
-          it('should add only one user fictional message without adding an assistant message for context, and the user message should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: true,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
+    context('when userId is not defined', function () {
+      it('returns true', function () {
+        const chat = domainBuilder.llm.buildChat({ userId: null });
 
-            // when
-            chat.addAttachmentContextMessages('winter_lyrics.txt', 'some user message along with the attachment');
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: true,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-        context('when attachment has been submitted alone', function () {
-          it('should add only one user fictional message without adding an assistant message for context, and the user message should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: true,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('winter_lyrics.txt', null);
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'winter_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: true,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-      });
-
-      context('when attachment is the wrong one', function () {
-        context('when attachment has been submitted along with a message', function () {
-          it('should add a user fictional message and not add any assistant context message, and should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: true,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('summer_lyrics.txt', 'some user message along with the attachment');
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'summer_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: true,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
-        context('when attachment has been submitted alone', function () {
-          it('should add a user fictional message and not add any assistant context message, and should not be sent to LLM', function () {
-            // given
-            const chat = new Chat({
-              id: 'some-chat-id',
-              configuration: new Configuration({
-                id: 'some-config-id',
-                attachment: {
-                  name: 'winter_lyrics.txt',
-                  context: "J'étais assise sur une pierre\nDes larmes coulaient sur mon visage",
-                },
-              }),
-              hasAttachmentContextBeenAdded: true,
-              messages: [
-                new Message({ index: 0, content: 'some message', isFromUser: true }),
-                new Message({ index: 1, content: 'some answer', isFromUser: false }),
-              ],
-            });
-
-            // when
-            chat.addAttachmentContextMessages('summer_lyrics.txt', null);
-
-            // then
-            const chatDTO = chat.toDTO();
-            expect(chatDTO).to.have.property('hasAttachmentContextBeenAdded', true);
-            expect(chatDTO).to.have.deep.property('messages', [
-              {
-                index: 0,
-                content: 'some message',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 1,
-                content: 'some answer',
-                attachmentName: undefined,
-                attachmentContext: undefined,
-                isFromUser: false,
-                shouldBeCountedAsAPrompt: false,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: false,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-              {
-                index: 2,
-                content: undefined,
-                attachmentName: 'summer_lyrics.txt',
-                attachmentContext: undefined,
-                isFromUser: true,
-                shouldBeCountedAsAPrompt: true,
-                shouldBeForwardedToLLM: false,
-                shouldBeRenderedInPreview: true,
-                hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-                haveVictoryConditionsBeenFulfilled: undefined,
-                wasModerated: undefined,
-                hasErrorOccurred: undefined,
-              },
-            ]);
-          });
-        });
+        expect(chat.isPreview).to.be.true;
       });
     });
   });
 
-  describe('#get currentPromptsCount', function () {
-    context('when chat has no messages at all', function () {
-      it('should return 0', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [],
-        });
-
-        // then
-        expect(chat).to.have.property('currentPromptsCount', 0);
-      });
-    });
-
-    context('when chat has no user messages', function () {
-      it('should return 0', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configurationId: 'some-config-id',
-          hasAttachmentContextBeenAdded: false,
-          messages: [
-            new Message({ index: 0, content: 'some message', isFromUser: false, shouldBeCountedAsAPrompt: true }),
-            new Message({
-              index: 1,
-              content: 'some other message',
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-            }),
-          ],
-        });
-
-        // then
-        expect(chat).to.have.property('currentPromptsCount', 0);
-      });
-    });
-
-    context('when chat has user messages', function () {
-      it('should return the number of user messages to count', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [
-            new Message({ index: 0, content: 'message llm 1', isFromUser: false, shouldBeCountedAsAPrompt: false }),
-            new Message({ index: 1, content: 'message user 1', isFromUser: true, shouldBeCountedAsAPrompt: true }),
-            new Message({ index: 2, content: 'message llm 2', isFromUser: false, shouldBeCountedAsAPrompt: false }),
-            new Message({ index: 3, content: 'message user 2', isFromUser: true, shouldBeCountedAsAPrompt: true }),
-          ],
-        });
-
-        // then
-        expect(chat).to.have.property('currentPromptsCount', 2);
-      });
-    });
-
-    context('when chat has user messages that should not be counted', function () {
-      it('should return the number of user messages to count', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({
-            id: 'some-config-id',
-            inputMaxChars: 500,
-            inputMaxPrompts: 4,
-            attachmentName: 'test.csv',
-            attachmentContext: 'le contexte',
+  describe('#get currentUserPromptsCount', function () {
+    context('when the expected attachment has already been added to context', function () {
+      let configuration, messages;
+      beforeEach(function () {
+        configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'recette_patate.pdf' });
+        messages = [
+          domainBuilder.llm.buildUserMessage({
+            index: 0,
+            content: 'G besoin de cb de patates',
+            attachmentName: 'recette_patate.pdf',
           }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [
-            new Message({ index: 0, content: 'message llm 1', isFromUser: false, shouldBeCountedAsAPrompt: false }),
-            new Message({ index: 1, content: 'message user 1', isFromUser: true, shouldBeCountedAsAPrompt: true }),
-            new Message({ index: 2, content: 'message llm 2', isFromUser: false, shouldBeCountedAsAPrompt: false }),
-            new Message({ index: 3, content: 'message user 2', isFromUser: true, shouldBeCountedAsAPrompt: false }),
-          ],
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: '3 !' }),
+        ];
+      });
+
+      context('after user sends a message with content but not attachment', function () {
+        it('should count', function () {
+          messages.push(domainBuilder.llm.buildUserMessage({ index: 2, content: 'wow, merci', attachmentName: null }));
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with content and wrong attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: 'wow, merci, et dans cette recette ?',
+              attachmentName: 'recette_haricots.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with content and right attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: 'wow, merci, et dans cette recette ?',
+              attachmentName: 'recette_patate.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with no content and wrong attachment', function () {
+        it('should NOT count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: null,
+              attachmentName: 'recette_chaussettes.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(1);
+        });
+      });
+
+      context('after user sends a message with no content and right attachment', function () {
+        it('should NOT count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: null,
+              attachmentName: 'recette_patate.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(1);
+        });
+      });
+    });
+
+    context('when the expected attachment has not already been added to context', function () {
+      let configuration, messages;
+      beforeEach(function () {
+        configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'recette_patate.pdf' });
+        messages = [
+          domainBuilder.llm.buildUserMessage({
+            index: 0,
+            content: 'G besoin de cb de patates',
+            attachmentName: null,
+          }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'Beaucoup' }),
+        ];
+      });
+
+      context('after user sends a message with content but not attachment', function () {
+        it('should count', function () {
+          messages.push(domainBuilder.llm.buildUserMessage({ index: 2, content: 'wow, merci', attachmentName: null }));
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with content and wrong attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: 'wow, merci, et dans cette recette ?',
+              attachmentName: 'recette_haricots.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with content and right attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: 'wow, merci, et dans cette recette ?',
+              attachmentName: 'recette_patate.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with no content and wrong attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: null,
+              attachmentName: 'recette_chaussettes.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+
+      context('after user sends a message with no content and right attachment', function () {
+        it('should count', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 2,
+              content: null,
+              attachmentName: 'recette_patate.pdf',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.currentUserPromptsCount).to.equal(2);
+        });
+      });
+    });
+  });
+
+  describe('#get messagesForInference', function () {
+    context('when the conversation only contains simple messages', function () {
+      it('should return a list of messages for inference', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage(),
+          domainBuilder.llm.buildAssistantMessage(),
+          domainBuilder.llm.buildUserMessage({ index: 2, content: 'super, merci chatpix' }),
+        ];
+        const chat = domainBuilder.llm.buildChat({ messages });
+
+        const messagesForInference = chat.messagesForInference;
+        expect(messagesForInference.length).to.equal(3);
+        expect(messagesForInference[0].content).to.equal(messages[0].content);
+        expect(messagesForInference[0].role).to.equal(messages[0].emitter);
+        expect(messagesForInference[1].content).to.equal(messages[1].content);
+        expect(messagesForInference[1].role).to.equal(messages[1].emitter);
+        expect(messagesForInference[2].content).to.equal(messages[2].content);
+        expect(messagesForInference[2].role).to.equal(messages[2].emitter);
+      });
+    });
+
+    context('when some user messages have been moderated', function () {
+      it('should return the list of messages not moderated', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage(),
+          domainBuilder.llm.buildAssistantMessage(),
+          domainBuilder.llm.buildUserMessage({
+            index: 2,
+            content: 'Je vais écrire un truc méchant!',
+            wasModerated: true,
+          }),
+          domainBuilder.llm.buildUserMessage({ index: 3, content: 'mais non je suis pipou' }),
+        ];
+        const chat = domainBuilder.llm.buildChat({ messages });
+
+        const messagesForInference = chat.messagesForInference;
+        expect(messagesForInference.length).to.equal(3);
+        expect(messagesForInference[0].content).to.equal(messages[0].content);
+        expect(messagesForInference[0].role).to.equal(messages[0].emitter);
+        expect(messagesForInference[1].content).to.equal(messages[1].content);
+        expect(messagesForInference[1].role).to.equal(messages[1].emitter);
+        expect(messagesForInference[2].content).to.equal(messages[3].content);
+        expect(messagesForInference[2].role).to.equal(messages[3].emitter);
+      });
+    });
+
+    context('when the expected attachment has been sent by the user', function () {
+      let configuration, messages, validAttachmentUserMessage, validAttachmentAssistantMessage;
+      beforeEach(function () {
+        configuration = domainBuilder.llm.buildConfiguration({
+          attachmentName: 'recette_patate.pdf',
+          attachmentContext: `Soupe de patates aux poireaux :
+          Ingrédients:
+          - 3 patates
+          - 2 poireaux
+          `,
+        });
+        messages = [
+          domainBuilder.llm.buildUserMessage({
+            index: 0,
+            attachmentName: 'recette_patate.pdf',
+            content: null,
+          }),
+        ];
+        validAttachmentUserMessage =
+          "<system_notification>L'utilisateur a téléversé une pièce jointe : <attachment_name>recette_patate.pdf</attachment_name></system_notification>";
+        validAttachmentAssistantMessage = `<read_attachment_tool>Lecture de la pièce jointe, recette_patate.pdf : <attachment_content>Soupe de patates aux poireaux :
+          Ingrédients:
+          - 3 patates
+          - 2 poireaux
+          </attachment_content></read_attachment_tool>`;
+      });
+
+      context('when the user sends a message with content only', function () {
+        it('should still have the attachment in context', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({ index: 1, content: 'Combien de patates ?', attachmentName: null }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          const messagesForInference = chat.messagesForInference;
+          expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+          expect(messagesForInference[0].role).to.equal('user');
+          expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+          expect(messagesForInference[1].role).to.equal('assistant');
+          expect(messagesForInference[2].content).to.equal('Combien de patates ?');
+          expect(messagesForInference[2].role).to.equal('user');
+          expect(messagesForInference.length).to.equal(3);
+        });
+      });
+
+      context('when the user sends a message with only wrong attachment', function () {
+        context('when the user sends another message afterwards', function () {
+          it('should not include the wrong attachment message', function () {
+            messages.push(
+              domainBuilder.llm.buildUserMessage({ index: 1, attachmentName: 'chocolat.txt', content: null }),
+              domainBuilder.llm.buildUserMessage({ index: 2, content: 'Combien de poireaux ?' }),
+            );
+            const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+            const messagesForInference = chat.messagesForInference;
+            expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+            expect(messagesForInference[0].role).to.equal('user');
+            expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+            expect(messagesForInference[1].role).to.equal('assistant');
+            expect(messagesForInference[2].content).to.equal('Combien de poireaux ?');
+            expect(messagesForInference[2].role).to.equal('user');
+            expect(messagesForInference.length).to.equal(3);
+          });
+        });
+      });
+
+      context('when the user sends a message with wrong attachment and content', function () {
+        it('should ignore the attachment and include the message content', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 1,
+              attachmentName: 'erreur.txt',
+              content: 'Et combien de grammes de sel ?',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          const messagesForInference = chat.messagesForInference;
+          expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+          expect(messagesForInference[0].role).to.equal('user');
+          expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+          expect(messagesForInference[1].role).to.equal('assistant');
+          expect(messagesForInference[2].content).to.equal('Et combien de grammes de sel ?');
+          expect(messagesForInference[2].role).to.equal('user');
+          expect(messagesForInference.length).to.equal(3);
+        });
+      });
+
+      context('when the user sends a message with only valid attachment', function () {
+        context('when the user sends another message afterwards', function () {
+          it('should not include a new valid attachment message', function () {
+            messages.push(
+              domainBuilder.llm.buildUserMessage({ index: 1, attachmentName: 'recette_patate.pdf', content: null }),
+              domainBuilder.llm.buildUserMessage({ index: 2, content: 'Combien de poireaux ?' }),
+            );
+            const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+            const messagesForInference = chat.messagesForInference;
+            expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+            expect(messagesForInference[0].role).to.equal('user');
+            expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+            expect(messagesForInference[1].role).to.equal('assistant');
+            expect(messagesForInference[2].content).to.equal('Combien de poireaux ?');
+            expect(messagesForInference[2].role).to.equal('user');
+            expect(messagesForInference.length).to.equal(3);
+          });
+        });
+      });
+
+      context('when the user sends a message with valid attachment and content', function () {
+        it('should ignore the attachment and include the message content', function () {
+          messages.push(
+            domainBuilder.llm.buildUserMessage({
+              index: 1,
+              attachmentName: 'recette_patate.pdf',
+              content: 'Et combien de grammes de sel ?',
+            }),
+          );
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          const messagesForInference = chat.messagesForInference;
+          expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+          expect(messagesForInference[0].role).to.equal('user');
+          expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+          expect(messagesForInference[1].role).to.equal('assistant');
+          expect(messagesForInference[2].content).to.equal('Et combien de grammes de sel ?');
+          expect(messagesForInference[2].role).to.equal('user');
+          expect(messagesForInference.length).to.equal(3);
+        });
+      });
+    });
+
+    context('when the expected attachment has not been sent by the user', function () {
+      let configuration, validAttachmentUserMessage, validAttachmentAssistantMessage;
+      beforeEach(function () {
+        configuration = domainBuilder.llm.buildConfiguration({
+          attachmentName: 'recette_patate.pdf',
+          attachmentContext: `Soupe de patates aux poireaux :
+          Ingrédients:
+          - 3 patates
+          - 2 poireaux
+          `,
+        });
+        validAttachmentUserMessage =
+          "<system_notification>L'utilisateur a téléversé une pièce jointe : <attachment_name>recette_patate.pdf</attachment_name></system_notification>";
+        validAttachmentAssistantMessage = `<read_attachment_tool>Lecture de la pièce jointe, recette_patate.pdf : <attachment_content>Soupe de patates aux poireaux :
+          Ingrédients:
+          - 3 patates
+          - 2 poireaux
+          </attachment_content></read_attachment_tool>`;
+      });
+
+      context('when the user sends a message with only wrong attachment', function () {
+        context('when the user sends another message afterwards', function () {
+          it('should not include the wrong attachment message', function () {
+            const messages = [
+              domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'chocolat.txt', content: null }),
+              domainBuilder.llm.buildUserMessage({ index: 1, content: 'Combien de poireaux ?' }),
+            ];
+            const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+            const messagesForInference = chat.messagesForInference;
+            expect(messagesForInference[0].content).to.equal('Combien de poireaux ?');
+            expect(messagesForInference[0].role).to.equal('user');
+            expect(messagesForInference.length).to.equal(1);
+          });
+        });
+      });
+
+      context('when the user sends a message with wrong attachment and content', function () {
+        context('when the user sends another message afterwards', function () {
+          it('should not include the wrong attachment message', function () {
+            const messages = [
+              domainBuilder.llm.buildUserMessage({
+                index: 0,
+                attachmentName: 'wrong.pdf',
+                content: 'Combien de patates ?',
+              }),
+              domainBuilder.llm.buildUserMessage({ index: 1, content: 'Ah mince' }),
+            ];
+            const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+            const messagesForInference = chat.messagesForInference;
+            expect(messagesForInference[0].content).to.equal('Ah mince');
+            expect(messagesForInference[0].role).to.equal('user');
+            expect(messagesForInference.length).to.equal(1);
+          });
+        });
+      });
+
+      context('when the user sends a message with only valid attachment', function () {
+        context('when the user sends another message afterwards', function () {
+          it('should include the valid attachment message', function () {
+            const messages = [
+              domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'recette_patate.pdf', content: null }),
+              domainBuilder.llm.buildUserMessage({ index: 1, content: 'Combien de poireaux ?' }),
+            ];
+            const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+            const messagesForInference = chat.messagesForInference;
+            expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+            expect(messagesForInference[0].role).to.equal('user');
+            expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+            expect(messagesForInference[1].role).to.equal('assistant');
+            expect(messagesForInference[2].content).to.equal('Combien de poireaux ?');
+            expect(messagesForInference[2].role).to.equal('user');
+            expect(messagesForInference.length).to.equal(3);
+          });
+        });
+      });
+
+      context('when the user sends a message with valid attachment and content', function () {
+        it('should include the attachment and message content', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({
+              attachmentName: 'recette_patate.pdf',
+              content: 'Combien de patates ?',
+            }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          const messagesForInference = chat.messagesForInference;
+          expect(messagesForInference[0].content).to.equal(validAttachmentUserMessage);
+          expect(messagesForInference[0].role).to.equal('user');
+          expect(messagesForInference[1].content).to.equal(validAttachmentAssistantMessage);
+          expect(messagesForInference[1].role).to.equal('assistant');
+          expect(messagesForInference[2].content).to.equal('Combien de patates ?');
+          expect(messagesForInference[2].role).to.equal('user');
+          expect(messagesForInference.length).to.equal(3);
+        });
+      });
+    });
+  });
+
+  describe('#get shouldSendForInference', function () {
+    const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'concentration.pdf' });
+
+    context('when attachment context is already loaded', function () {
+      context('when last user message only has a content', function () {
+        it('should return true', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'concentration.pdf', content: null }),
+            domainBuilder.llm.buildUserMessage({ index: 1, content: 'On mange quoi ?' }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(true);
+        });
+      });
+
+      context('when last user message has a content and a valid attachment', function () {
+        it('should return true', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'concentration.pdf', content: null }),
+            domainBuilder.llm.buildUserMessage({
+              index: 1,
+              attachmentName: 'concentration.pdf',
+              content: 'On mange quoi ?',
+            }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(true);
+        });
+      });
+
+      context('when last user message has no content and a valid attachment', function () {
+        it('should return false', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'concentration.pdf', content: null }),
+            domainBuilder.llm.buildUserMessage({ index: 1, attachmentName: 'concentration.pdf', content: null }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(false);
+        });
+      });
+
+      context('when last user message has a content and a wrong attachment', function () {
+        it('should return true', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'concentration.pdf', content: null }),
+            domainBuilder.llm.buildUserMessage({ index: 1, attachmentName: 'nul.pdf', content: 'On mange quoi ?' }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(true);
+        });
+      });
+
+      context('when last user message has no content and a wrong attachment', function () {
+        it('should return false', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ index: 0, attachmentName: 'concentration.pdf', content: null }),
+            domainBuilder.llm.buildUserMessage({ index: 1, attachmentName: 'menu_gouter.pdf', content: null }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(false);
+        });
+      });
+    });
+
+    context('when attachment context is not already loaded', function () {
+      context('when last user message only has a content', function () {
+        it('should return true', function () {
+          const messages = [domainBuilder.llm.buildUserMessage({ content: 'On mange quoi ?' })];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(true);
+        });
+      });
+
+      context('when last user message has a content and a valid attachment', function () {
+        it('should return true', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ attachmentName: 'concentration.pdf', content: 'On mange quoi ?' }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(true);
+        });
+      });
+
+      context('when last user message has no content and a valid attachment', function () {
+        it('should return false', function () {
+          const messages = [domainBuilder.llm.buildUserMessage({ attachmentName: 'concentration.pdf', content: null })];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(false);
+        });
+      });
+
+      context('when last user message has a content and a wrong attachment', function () {
+        it('should return false', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ attachmentName: 'nul.pdf', content: 'On mange quoi ?' }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(false);
+        });
+      });
+
+      context('when last user message has no content and a wrong attachment', function () {
+        it('should return false', function () {
+          const messages = [domainBuilder.llm.buildUserMessage({ attachmentName: 'nul.pdf', content: null })];
+          const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+          expect(chat.shouldSendForInference).to.equal(false);
+        });
+      });
+    });
+
+    context('when there are no messages', function () {
+      it('should return false', function () {
+        const chat = domainBuilder.llm.buildChat({ configuration, messages: [] });
+
+        expect(chat.shouldSendForInference).to.equal(false);
+      });
+    });
+
+    context('when the last message is a user message with a content and no attachment', function () {
+      it('should return true', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: "C'est l'heure de s'arrêter pour le goûter" })];
+        const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+        expect(chat.shouldSendForInference).to.equal(true);
+      });
+    });
+  });
+
+  describe('#get lastAttachmentStatus', function () {
+    const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'test.pdf' });
+
+    context('when last user message has a valid attachment', function () {
+      it('should return SUCCESS', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: 'Ok', attachmentName: 'test.pdf' })];
+        const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+        expect(chat.lastAttachmentStatus).to.equal(Chat.ATTACHMENT_STATUS.SUCCESS);
+      });
+    });
+
+    context('when last user message has a wrong attachment', function () {
+      it('should return FAILURE', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: 'Pas ok', attachmentName: 'nul.txt' })];
+        const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+        expect(chat.lastAttachmentStatus).to.equal(Chat.ATTACHMENT_STATUS.FAILURE);
+      });
+    });
+
+    context('when last user message has no attachment', function () {
+      it('should return NONE', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: 'Salut' })];
+        const chat = domainBuilder.llm.buildChat({ configuration, messages });
+
+        expect(chat.lastAttachmentStatus).to.equal(Chat.ATTACHMENT_STATUS.NONE);
+      });
+    });
+
+    context('when no attachment is expected', function () {
+      it('should return NONE', function () {
+        const chat = domainBuilder.llm.buildChat({
+          messages: [domainBuilder.llm.buildUserMessage({ content: 'Salut ' })],
         });
 
-        // then
-        expect(chat).to.have.property('currentPromptsCount', 1);
+        expect(chat.lastAttachmentStatus).to.equal(Chat.ATTACHMENT_STATUS.NONE);
+      });
+    });
+  });
+
+  describe('#get lastUserMessage', function () {
+    context('when there are messages', function () {
+      it('should get the last user message', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Salut' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'Bonjour !' }),
+          domainBuilder.llm.buildUserMessage({ index: 2, content: 'Ça va ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 3, content: 'Plus maintenant.' }),
+          domainBuilder.llm.buildUserMessage({ index: 4, content: 'Méchant va !' }),
+        ];
+        const chat = domainBuilder.llm.buildChat({ messages });
+
+        const lastUserMessage = chat.lastUserMessage;
+
+        expect(lastUserMessage).to.deepEqualInstance(
+          domainBuilder.llm.buildUserMessage({ index: 4, content: 'Méchant va !' }),
+        );
+      });
+
+      context('when last message is an assistant message', function () {
+        it('should return last USER message', function () {
+          const chat = domainBuilder.llm.buildChat({
+            messages: [
+              domainBuilder.llm.buildUserMessage({ index: 0, content: 'Salut' }),
+              domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'Bonjour !' }),
+            ],
+          });
+
+          const lastUserMessage = chat.lastUserMessage;
+
+          expect(lastUserMessage).to.deepEqualInstance(
+            domainBuilder.llm.buildUserMessage({ index: 0, content: 'Salut' }),
+          );
+        });
+      });
+    });
+
+    context('when there are no messages', function () {
+      it('should return undefined', function () {
+        const chat = domainBuilder.llm.buildChat({ messages: [] });
+
+        expect(chat.lastUserMessage).to.be.undefined;
+      });
+    });
+  });
+
+  describe('#addUserMessage', function () {
+    context('when neither content nor attachmentName are provided', function () {
+      it('throws', function () {
+        const chat = domainBuilder.llm.buildChat();
+        expect(() => chat.addUserMessage(null, null)).to.throw(NoAttachmentNorMessageProvidedError);
+      });
+    });
+
+    context('when adding an user message with an attachmentName', function () {
+      context('no attachment is expected in the configuration', function () {
+        it('throws', function () {
+          const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: null });
+          const chat = domainBuilder.llm.buildChat({ configuration });
+          expect(() => chat.addUserMessage(null, 'foo.txt')).to.throw(NoAttachmentNeededError);
+        });
+      });
+
+      context('an attachment is expected in the configuration', function () {
+        it('adds the message', function () {
+          const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'foo.txt' });
+          const chat = domainBuilder.llm.buildChat({ configuration });
+
+          chat.addUserMessage('patate', 'foo.txt');
+
+          expect(chat.messages[0]).to.deepEqualInstance(
+            new Message({
+              index: 0,
+              content: 'patate',
+              attachmentName: 'foo.txt',
+              emitter: 'user',
+              wasModerated: null,
+            }),
+          );
+        });
+      });
+    });
+
+    context('when message content exceeds configuration inputMaxChars', function () {
+      it('throws', function () {
+        const configuration = domainBuilder.llm.buildConfiguration({ inputMaxChars: 10 });
+        const chat = domainBuilder.llm.buildChat({ configuration });
+
+        expect(() => chat.addUserMessage('Mon tonton tond ton tonton et ton tonton est tout tondu', null)).to.throw(
+          TooLargeMessageInputError,
+        );
+      });
+    });
+
+    context('when user message count exceeds configuration inputMaxPrompts', function () {
+      it('throws', function () {
+        const configuration = domainBuilder.llm.buildConfiguration({ inputMaxPrompts: 1 });
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Peux tu m aider ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'oui bien sur' }),
+        ];
+        const chat = domainBuilder.llm.buildChat({ messages, configuration });
+
+        expect(() => chat.addUserMessage('Mon tonton tond ton tonton et ton tonton est tout tondu', null)).to.throw(
+          MaxPromptsReachedError,
+        );
+      });
+    });
+
+    context('when chat already contains some messages', function () {
+      context('when previous message is already a user message', function () {
+        context('when previous message is a message with content only', function () {
+          it('throws', function () {
+            const messages = [domainBuilder.llm.buildUserMessage({ content: 'Peux tu m aider ?' })];
+            const chat = domainBuilder.llm.buildChat({ messages });
+
+            expect(() => chat.addUserMessage('Peux tu m aider ?')).to.throw(IncorrectMessagesOrderingError);
+          });
+        });
+
+        context('when previous message is a message with an attachment', function () {
+          it('should add the new user message', function () {
+            const messages = [domainBuilder.llm.buildUserMessage({ attachmentName: 'aide.txt', content: null })];
+            const configuration = domainBuilder.llm.buildConfiguration({ attachmentName: 'aide.txt' });
+            const chat = domainBuilder.llm.buildChat({ messages, configuration });
+
+            chat.addUserMessage("Peux-tu m'aider ?");
+
+            expect(chat.messages[0]).to.deepEqualInstance(
+              new Message({
+                index: 0,
+                content: null,
+                attachmentName: 'aide.txt',
+                emitter: 'user',
+                wasModerated: false,
+              }),
+            );
+            expect(chat.messages[1]).to.deepEqualInstance(
+              new Message({
+                index: 1,
+                content: "Peux-tu m'aider ?",
+                attachmentName: null,
+                emitter: 'user',
+                wasModerated: null,
+              }),
+            );
+            expect(chat.messages.length).to.equal(2);
+          });
+        });
+
+        context('when previous message was moderated', function () {
+          it('should add the new user message', function () {
+            const messages = [
+              domainBuilder.llm.buildUserMessage({ content: 'Ordure, peux tu m aider ?', wasModerated: true }),
+            ];
+            const chat = domainBuilder.llm.buildChat({ messages });
+            chat.addUserMessage("Peux-tu m'aider ?");
+
+            expect(chat.messages.length).to.equal(2);
+            expect(chat.messages[1]).to.deepEqualInstance(
+              new Message({
+                index: 1,
+                content: "Peux-tu m'aider ?",
+                attachmentName: null,
+                emitter: 'user',
+                wasModerated: null,
+              }),
+            );
+          });
+        });
+      });
+
+      it('should add user message to the messages and increment index attribute accordingly', function () {
+        const messages = [
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'Peux tu m aider ?' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'oui bien sur' }),
+        ];
+        const chat = domainBuilder.llm.buildChat({ messages });
+        chat.addUserMessage('patate');
+
+        expect(chat.messages[2]).to.deepEqualInstance(
+          new Message({
+            index: 2,
+            content: 'patate',
+            attachmentName: null,
+            emitter: 'user',
+            wasModerated: null,
+          }),
+        );
+      });
+    });
+  });
+
+  describe('#addAssistantMessage', function () {
+    context('when message has content', function () {
+      context('when chat has no messages', function () {
+        it('throws', function () {
+          const chat = domainBuilder.llm.buildChat();
+
+          expect(() => chat.addAssistantMessage('chocolat')).to.throw(IncorrectMessagesOrderingError);
+        });
+      });
+
+      context('when previous message is already an assistant message', function () {
+        it('throws', function () {
+          const messages = [
+            domainBuilder.llm.buildUserMessage({ content: 'Parle avec toi-même pour voir' }),
+            domainBuilder.llm.buildAssistantMessage({ content: 'Très bien. Salut ChatPix, ça va ?' }),
+          ];
+          const chat = domainBuilder.llm.buildChat({ messages });
+
+          expect(() => chat.addAssistantMessage('Oui et toi ?')).to.throw(IncorrectMessagesOrderingError);
+        });
+      });
+
+      it('adds the message', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: '1 + 1 = 3 😌' })];
+        const chat = domainBuilder.llm.buildChat({ messages });
+
+        chat.addAssistantMessage(
+          'Révolutionnaire. Toutes mes félicitations pour ce travail de recherche exceptionnel.',
+        );
+
+        expect(chat.messages[1]).to.deepEqualInstance(
+          new Message({
+            index: 1,
+            content: 'Révolutionnaire. Toutes mes félicitations pour ce travail de recherche exceptionnel.',
+            attachmentName: null,
+            emitter: 'assistant',
+            wasModerated: null,
+          }),
+        );
+      });
+    });
+
+    context('when message has no content', function () {
+      it('throws', function () {
+        const messages = [domainBuilder.llm.buildUserMessage({ content: 'Ça donne quoi " ".trim() en javascript ?' })];
+        const chat = domainBuilder.llm.buildChat({ messages });
+
+        expect(() => chat.addAssistantMessage(null)).to.throw(NoAttachmentNorMessageProvidedError);
+      });
+    });
+  });
+
+  describe('#updateTokenConsumption', function () {
+    it('adds to corresponding total the token counts in param', function () {
+      // given
+      const configurationDTO = Symbol('configurationDTO');
+      const chat = new Chat({
+        id: 'some-chat-id',
+        userId: 123,
+        configurationId: 'abc123',
+        configuration: new Configuration(configurationDTO),
+        haveVictoryConditionsBeenFulfilled: false,
+        totalInputTokens: 1_000,
+        totalOutputTokens: 10_000,
+        messages: [
+          new Message({
+            index: 0,
+            content: 'message user 1',
+            emitter: 'user',
+          }),
+          new Message({
+            index: 1,
+            content: 'message llm 1',
+            emitter: 'assistant',
+          }),
+        ],
+      });
+
+      // when
+      chat.updateTokenConsumption(2_000, 5_000);
+
+      // then
+      expect(chat.toDTO()).to.deep.equal({
+        id: 'some-chat-id',
+        userId: 123,
+        assessmentId: undefined,
+        challengeId: undefined,
+        moduleId: undefined,
+        passageId: undefined,
+        configurationId: 'abc123',
+        configuration: configurationDTO,
+        haveVictoryConditionsBeenFulfilled: false,
+        totalInputTokens: 3_000,
+        totalOutputTokens: 15_000,
+        messages: [
+          {
+            index: 0,
+            content: 'message user 1',
+            attachmentName: undefined,
+            emitter: 'user',
+            wasModerated: undefined,
+          },
+          {
+            index: 1,
+            content: 'message llm 1',
+            attachmentName: undefined,
+            emitter: 'assistant',
+            wasModerated: undefined,
+          },
+        ],
       });
     });
   });
@@ -1152,7 +1056,6 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
         const chat = new Chat({
           id: 'some-chat-id',
           configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
           messages: [],
         });
 
@@ -1170,7 +1073,6 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
         chat = new Chat({
           id: 'some-chat-id',
           configuration: new Configuration({ id: 'some-config-id', attachment: { name: 'Le plan de ma maison.txt' } }),
-          hasAttachmentContextBeenAdded: false,
           messages: [],
         });
       });
@@ -1235,322 +1137,24 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
     });
   });
 
-  describe('#get messagesToForwardToLLM', function () {
-    context('when chat has no messages at all', function () {
-      it('should return an empty array', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({ id: 'some-config-id' }),
-          hasAttachmentContextBeenAdded: false,
-          messages: [],
-        });
-
-        // then
-        expect(chat.messagesToForwardToLLM).to.deep.equal([]);
-      });
-    });
-    context('attachments', function () {
-      it('returns well formatted attachment messages', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({
-            id: 'some-config-id',
-            attachment: { name: 'file.txt', context: "Ceci n'est pas une pipe." },
-          }),
-          hasAttachmentContextBeenAdded: true,
-          messages: [
-            new Message({ index: 0, attachmentName: 'file.txt', isFromUser: true, shouldBeForwardedToLLM: true }),
-            new Message({
-              index: 1,
-              attachmentName: 'file.txt',
-              attachmentContext: "Ceci n'est pas une pipe.",
-              isFromUser: false,
-              shouldBeForwardedToLLM: true,
-            }),
-            new Message({
-              index: 2,
-              content: 'Quel instrument pour fumer est mentionné dans mon fichier ?',
-              isFromUser: true,
-              shouldBeForwardedToLLM: true,
-            }),
-          ],
-        });
-
-        // then
-        expect(chat.messagesToForwardToLLM).to.deep.equal([
-          {
-            content:
-              "<system_notification>L'utilisateur a téléversé une pièce jointe : <attachment_name>file.txt</attachment_name></system_notification>",
-            role: 'user',
-          },
-          {
-            content:
-              "<read_attachment_tool>Lecture de la pièce jointe, file.txt : <attachment_content>Ceci n'est pas une pipe.</attachment_content></read_attachment_tool>",
-            role: 'assistant',
-          },
-          {
-            content: 'Quel instrument pour fumer est mentionné dans mon fichier ?',
-            role: 'user',
-          },
-        ]);
-      });
-    });
-    context('messages not to forward', function () {
-      it('returns messages except the ones to not forward', function () {
-        // given
-        const chat = new Chat({
-          id: 'some-chat-id',
-          configuration: new Configuration({
-            id: 'some-config-id',
-            attachment: { name: 'file.txt', context: "Ceci n'est pas une pipe." },
-          }),
-          hasAttachmentContextBeenAdded: true,
-          messages: [
-            new Message({ index: 0, content: 'message user 1', isFromUser: true, shouldBeForwardedToLLM: true }),
-            new Message({ index: 1, content: 'message assistant 1', isFromUser: false, shouldBeForwardedToLLM: true }),
-            new Message({
-              index: 2,
-              content: 'message user à ignorer',
-              isFromUser: true,
-              shouldBeForwardedToLLM: false,
-            }),
-          ],
-        });
-
-        // then
-        expect(chat.messagesToForwardToLLM).to.deep.equal([
-          {
-            content: 'message user 1',
-            role: 'user',
-          },
-          {
-            content: 'message assistant 1',
-            role: 'assistant',
-          },
-        ]);
-      });
-    });
-  });
-
-  describe('#updateTokenConsumption', function () {
-    context('when chat has not initialized integer values for totalTokens attributes (old chats)', function () {
-      it('does nothing', function () {
-        // given
-        const configurationDTO = Symbol('configurationDTO');
-        const chat = new Chat({
-          id: 'some-chat-id',
-          userId: 123,
-          configurationId: 'abc123',
-          configuration: new Configuration(configurationDTO),
-          hasAttachmentContextBeenAdded: false,
-          messages: [
-            new Message({
-              index: 0,
-              content: 'message user 1',
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
-            }),
-            new Message({
-              index: 1,
-              content: 'message llm 1',
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeRenderedInPreview: true,
-              shouldBeForwardedToLLM: true,
-            }),
-          ],
-        });
-
-        // when
-        chat.updateTokenConsumption(2_000, 5_000);
-
-        // then
-        expect(chat.toDTO()).to.deep.equal({
-          id: 'some-chat-id',
-          userId: 123,
-          assessmentId: undefined,
-          challengeId: undefined,
-          moduleId: undefined,
-          passageId: undefined,
-          configurationId: 'abc123',
-          configuration: configurationDTO,
-          hasAttachmentContextBeenAdded: false,
-          totalInputTokens: undefined,
-          totalOutputTokens: undefined,
-          messages: [
-            {
-              index: 0,
-              content: 'message user 1',
-              attachmentName: undefined,
-              attachmentContext: undefined,
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
-              hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-              haveVictoryConditionsBeenFulfilled: undefined,
-              wasModerated: undefined,
-              hasErrorOccurred: undefined,
-            },
-            {
-              index: 1,
-              content: 'message llm 1',
-              attachmentName: undefined,
-              attachmentContext: undefined,
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeRenderedInPreview: true,
-              shouldBeForwardedToLLM: true,
-              hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-              haveVictoryConditionsBeenFulfilled: undefined,
-              wasModerated: undefined,
-              hasErrorOccurred: undefined,
-            },
-          ],
-        });
-      });
-    });
-
-    context('when chat has integer values for totalTokens attributes', function () {
-      it('adds to corresponding total the token counts in param', function () {
-        // given
-        const configurationDTO = Symbol('configurationDTO');
-        const chat = new Chat({
-          id: 'some-chat-id',
-          userId: 123,
-          configurationId: 'abc123',
-          configuration: new Configuration(configurationDTO),
-          hasAttachmentContextBeenAdded: false,
-          totalInputTokens: 1_000,
-          totalOutputTokens: 10_000,
-          messages: [
-            new Message({
-              index: 0,
-              content: 'message user 1',
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
-            }),
-            new Message({
-              index: 1,
-              content: 'message llm 1',
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeRenderedInPreview: true,
-              shouldBeForwardedToLLM: true,
-            }),
-          ],
-        });
-
-        // when
-        chat.updateTokenConsumption(2_000, 5_000);
-
-        // then
-        expect(chat.toDTO()).to.deep.equal({
-          id: 'some-chat-id',
-          userId: 123,
-          assessmentId: undefined,
-          challengeId: undefined,
-          moduleId: undefined,
-          passageId: undefined,
-          configurationId: 'abc123',
-          configuration: configurationDTO,
-          hasAttachmentContextBeenAdded: false,
-          totalInputTokens: 3_000,
-          totalOutputTokens: 15_000,
-          messages: [
-            {
-              index: 0,
-              content: 'message user 1',
-              attachmentName: undefined,
-              attachmentContext: undefined,
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
-              hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-              haveVictoryConditionsBeenFulfilled: undefined,
-              wasModerated: undefined,
-              hasErrorOccurred: undefined,
-            },
-            {
-              index: 1,
-              content: 'message llm 1',
-              attachmentName: undefined,
-              attachmentContext: undefined,
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeRenderedInPreview: true,
-              shouldBeForwardedToLLM: true,
-              hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-              haveVictoryConditionsBeenFulfilled: undefined,
-              wasModerated: undefined,
-              hasErrorOccurred: undefined,
-            },
-          ],
-        });
-      });
-    });
-  });
-
   describe('#toDTO', function () {
     it('should return the DTO version of the Chat model', function () {
       // given
-      const configurationDTO = Symbol('configurationDTO');
-      const chat = new Chat({
+      const chat = domainBuilder.llm.buildChat({
         id: 'some-chat-id',
         userId: 123,
         assessmentId: 456,
         challengeId: '789',
-        moduleId: 1234,
+        haveVictoryConditionsBeenFulfilled: false,
         configurationId: 'abc123',
-        configuration: new Configuration(configurationDTO),
-        hasAttachmentContextBeenAdded: true,
+        configuration: domainBuilder.llm.buildConfiguration(),
         totalInputTokens: 2_000,
         totalOutputTokens: 5_000,
         messages: [
-          new Message({
-            index: 0,
-            content: 'message user 1',
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            wasModerated: true,
-          }),
-          new Message({
-            index: 1,
-            content: 'message llm 1',
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeRenderedInPreview: true,
-            shouldBeForwardedToLLM: true,
-            hasErrorOccurred: true,
-          }),
-          new Message({
-            index: 2,
-            attachmentName: 'file.txt',
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-            haveVictoryConditionsBeenFulfilled: true,
-          }),
-          new Message({
-            index: 3,
-            attachmentName: 'file.txt',
-            attachmentContext: 'je suis un poulet',
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: false,
-          }),
+          domainBuilder.llm.buildUserMessage({ index: 0, content: 'message user 1', wasModerated: true }),
+          domainBuilder.llm.buildAssistantMessage({ index: 1, content: 'message llm 1' }),
+          domainBuilder.llm.buildUserMessage({ index: 2, content: 'message user 2', attachmentName: 'file.txt' }),
+          domainBuilder.llm.buildAssistantMessage({ index: 3, content: 'message llm 2' }),
         ],
       });
 
@@ -1563,69 +1167,41 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
         userId: 123,
         assessmentId: 456,
         challengeId: '789',
-        moduleId: 1234,
-        passageId: undefined,
+        moduleId: null,
+        passageId: null,
         configurationId: 'abc123',
-        configuration: configurationDTO,
-        hasAttachmentContextBeenAdded: true,
+        configuration: chat.configuration.toDTO(),
+        haveVictoryConditionsBeenFulfilled: false,
         totalInputTokens: 2_000,
         totalOutputTokens: 5_000,
         messages: [
           {
             index: 0,
             content: 'message user 1',
-            attachmentName: undefined,
-            attachmentContext: undefined,
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            haveVictoryConditionsBeenFulfilled: undefined,
+            emitter: 'user',
             wasModerated: true,
-            hasErrorOccurred: undefined,
+            attachmentName: null,
           },
           {
             index: 1,
             content: 'message llm 1',
-            attachmentName: undefined,
-            attachmentContext: undefined,
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeRenderedInPreview: true,
-            shouldBeForwardedToLLM: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            haveVictoryConditionsBeenFulfilled: undefined,
-            wasModerated: undefined,
-            hasErrorOccurred: true,
+            emitter: 'assistant',
+            attachmentName: null,
+            wasModerated: null,
           },
           {
             index: 2,
-            content: undefined,
+            content: 'message user 2',
+            emitter: 'user',
             attachmentName: 'file.txt',
-            attachmentContext: undefined,
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-            haveVictoryConditionsBeenFulfilled: true,
-            wasModerated: undefined,
-            hasErrorOccurred: undefined,
+            wasModerated: false,
           },
           {
             index: 3,
-            content: undefined,
-            attachmentName: 'file.txt',
-            attachmentContext: 'je suis un poulet',
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: false,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            haveVictoryConditionsBeenFulfilled: undefined,
-            wasModerated: undefined,
-            hasErrorOccurred: undefined,
+            content: 'message llm 2',
+            emitter: 'assistant',
+            attachmentName: null,
+            wasModerated: null,
           },
         ],
       });
@@ -1640,58 +1216,29 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
         userId: 123,
         configurationId: 'abc123',
         configuration: {},
-        hasAttachmentContextBeenAdded: true,
+        haveVictoryConditionsBeenFulfilled: false,
         messages: [
           {
             index: 0,
             content: 'message user 1',
-            attachmentName: undefined,
-            attachmentContext: undefined,
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            haveVictoryConditionsBeenFulfilled: undefined,
+            emitter: 'user',
             wasModerated: true,
           },
           {
             index: 1,
             content: 'message llm 1',
-            attachmentName: undefined,
-            attachmentContext: undefined,
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeRenderedInPreview: true,
-            shouldBeForwardedToLLM: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            haveVictoryConditionsBeenFulfilled: undefined,
-            wasModerated: undefined,
+            emitter: 'assistant',
           },
           {
             index: 2,
-            content: undefined,
+            content: 'message user 2',
+            emitter: 'user',
             attachmentName: 'file.txt',
-            attachmentContext: undefined,
-            isFromUser: true,
-            shouldBeCountedAsAPrompt: true,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: true,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-            haveVictoryConditionsBeenFulfilled: true,
-            wasModerated: undefined,
           },
           {
             index: 3,
-            content: undefined,
-            attachmentName: 'file.txt',
-            attachmentContext: 'je suis un poulet',
-            isFromUser: false,
-            shouldBeCountedAsAPrompt: false,
-            shouldBeForwardedToLLM: true,
-            shouldBeRenderedInPreview: false,
-            hasAttachmentBeenSubmittedAlongWithAPrompt: undefined,
-            wasModerated: undefined,
+            content: 'message llm 2',
+            emitter: 'assistant',
           },
         ],
       };
@@ -1706,43 +1253,29 @@ describe('LLM | Unit | Domain | Models | Chat', function () {
           userId: 123,
           configurationId: 'abc123',
           configuration: new Configuration({}), // Configuration model has no enumerable properties
-          hasAttachmentContextBeenAdded: true,
+          haveVictoryConditionsBeenFulfilled: false,
           messages: [
             new Message({
               index: 0,
               content: 'message user 1',
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
+              emitter: 'user',
               wasModerated: true,
             }),
             new Message({
               index: 1,
               content: 'message llm 1',
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeRenderedInPreview: true,
-              shouldBeForwardedToLLM: true,
+              emitter: 'assistant',
             }),
             new Message({
               index: 2,
               attachmentName: 'file.txt',
-              isFromUser: true,
-              shouldBeCountedAsAPrompt: true,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: true,
-              hasAttachmentBeenSubmittedAlongWithAPrompt: false,
-              haveVictoryConditionsBeenFulfilled: true,
+              content: 'message user 2',
+              emitter: 'user',
             }),
             new Message({
               index: 3,
-              attachmentName: 'file.txt',
-              attachmentContext: 'je suis un poulet',
-              isFromUser: false,
-              shouldBeCountedAsAPrompt: false,
-              shouldBeForwardedToLLM: true,
-              shouldBeRenderedInPreview: false,
+              content: 'message llm 2',
+              emitter: 'assistant',
             }),
           ],
         }),
