+++ api/tests/tooling/domain-builder/factory/llm/build-message.js
@@ -1,12 +1,12 @@
-import { MessageV2 } from '../../../../../src/llm/domain/models/ChatV2.js';
+import { Message } from '../../../../../src/llm/domain/models/Chat.js';
 
-export function buildUserMessageV2({
+export function buildUserMessage({
   index = 0,
   content = 'la capitale de la finlande stp',
   attachmentName = null,
   wasModerated = false,
 } = {}) {
-  return new MessageV2({
+  return new Message({
     index,
     content,
     attachmentName,
@@ -15,8 +15,8 @@ export function buildUserMessageV2({
   });
 }
 
-export function buildAssistantMessageV2({ index = 1, content = 'Helsinki bien sûr !!' } = {}) {
-  return new MessageV2({
+export function buildAssistantMessage({ index = 1, content = 'Helsinki bien sûr !!' } = {}) {
+  return new Message({
     index,
     content,
     attachmentName: null,
